+++ pioreactor/cluster_management/__init__.py
@@ -44,11 +44,11 @@ def get_active_workers_in_experiment(experiment: str) -> tuple[str, ...]:
 @click.command(name="add", short_help="add a pioreactor worker")
 @click.argument("hostname")
 @click.option("--password", "-p", default="raspberry")
-@click.option("--model-version", "-v", required=True, type=click.STRING)
 @click.option("--model-name", "-m", type=click.Choice(["pioreactor_20ml", "pioreactor_40ml"]), required=True)
+@click.option("--model-version", "-v", required=True, type=click.Choice(["1.0", "1.1"]))
 @click.option("--address", "-a")
 def add_worker(
-    hostname: str, password: str, model_version: str, model_name: str, address: str | None
+    hostname: str, password: str, model_name: str, model_version: str, address: str | None
 ) -> None:
     """
     Add a new pioreactor worker to the cluster. The pioreactor should already have the worker image installed and is turned on.